add: 完成mybatis-plus接口添加

+++ src/main/java/com/sinnevy/minproject/controller/UserController.java
@@ -63,7 +63,7 @@ public class UserController {
     }
 
     @GetMapping("/delete")
-    public ResponseEntity<Result> delete(Integer id) {
+    public ResponseEntity<Result> delete(Long id) {
         userService.delete(id);
         return ResponseEntity.ok(Result.success(null));
     }
